Use the shared studio controls in the sound settings panel
The UI-sound popover rendered raw native <input type="range">, checkboxes
and a bare <select>, so it looked nothing like the rest of the studio. Wire
it up to the same custom chrome every other control uses:

- Extract the presentational, fully-controlled SliderInput out of
  ControlledSliderInput (the Blender/Leva drag bar + tap-to-type). The RHF
  wrapper now just binds field value/onChange to it — behaviour-preserving.
- Drive the sound panel's sliders through SliderInput, its booleans through
  ToggleSwitch, and the preset picker through the segmented select bar
  (BarLabelSegment + invisible native select + chevron).

The panel now matches the studio's look and feel while still being driven by
the UI-sound store rather than the sketch form.
